35 DE PUNCTEEE

+++ functii_temaaaa.h
@@ -39,4 +39,4 @@ void scoatere_echipe(Echipa **, int , FILE *);
 
 void aflare_minim(Echipa *, double *, FILE *);
 
-void scoatere_echipa(Echipa **, int , FILE *);
+void scoatere_echipa(Echipa **, double , FILE *);
